+++ media/locales/de/messages.js
@@ -53,6 +53,15 @@ window.languageManager.loadMessages('de', {
 	CATEGORY_FUNCTIONS: 'Funktionen',
 	CATEGORY_ARDUINO: 'Arduino',
 	CATEGORY_SENSORS: 'Sensoren',
+	CATEGORY_MOTORS: 'Motoren',
+
+	// Servo Block Labels
+	SERVO_SETUP: 'Servomotor einrichten',
+	SERVO_PIN: 'Pin',
+	SERVO_SETUP_TOOLTIP: 'Servomotor-Variable deklarieren und Pin festlegen',
+	SERVO_MOVE: 'Servomotor drehen',
+	SERVO_ANGLE: 'Winkel',
+	SERVO_MOVE_TOOLTIP: 'Servomotor auf bestimmten Winkel drehen',
 
 	// Arduino Block Labels
 	ARDUINO_SETUP: 'Setup',