+++ media/locales/ko/messages.js
@@ -53,6 +53,15 @@ window.languageManager.loadMessages('ko', {
 	CATEGORY_FUNCTIONS: '함수',
 	CATEGORY_ARDUINO: 'Arduino',
 	CATEGORY_SENSORS: '센서',
+	CATEGORY_MOTORS: '모터',
+
+	// Servo Block Labels
+	SERVO_SETUP: '서보 모터 설정',
+	SERVO_PIN: '핀',
+	SERVO_SETUP_TOOLTIP: '서보 모터 변수를 선언하고 핀 설정',
+	SERVO_MOVE: '서보 모터 회전',
+	SERVO_ANGLE: '각도',
+	SERVO_MOVE_TOOLTIP: '서보 모터를 특정 각도로 회전',
 
 	// Arduino Block Labels
 	ARDUINO_SETUP: '설정',